Fix error reporting for duplicate plugin application
Previously, when we detected a duplicate plugin application, we would
always report the location of the first plugin applied, regardless of which
plugin that was. This led to nonsensical errors about a plugin being applied
at a location that pointed to a completely different plugin.

Now, we report the location of the first application of the plugin that was
duplicated.